requirements.txt
pytest>=8.0.0
allure-pytest>=2.13.5
playwright>=1.43.0